+++ packages/react-native-payments/jest.config.js
@@ -1,5 +1,6 @@
 module.exports = {
     ...require('../../get-jest.config.js')('react-native-payments'),
+    collectCoverageFrom: ['src/**/*.ts', '!src/**/*.spec.ts', '!src/**/*.web.ts'],
     coverageThreshold: {
         global: {
             branches: 100,