+++ packages/react-native-payments/src/class/payment-request/payment-request.web.ts
@@ -1,1 +1,6 @@
-export const { PaymentRequest } = typeof window === 'undefined' ? { PaymentRequest: null } : window;
+import type { WebPaymentRequestConstructor } from '../../type/web-payment-request-constructor.type';
+import type { Maybe } from '@rnw-community/shared';
+
+
+export const { PaymentRequest }: { PaymentRequest: Maybe<WebPaymentRequestConstructor> | undefined } =
+    typeof window === 'undefined' ? { PaymentRequest: null } : window;