+++ packages/react-native-payments/src/class/payment-response/payment-response.web.ts
@@ -1,1 +1,6 @@
-export const { PaymentResponse } = typeof window === 'undefined' ? { PaymentResponse: null } : window;
+import type { WebPaymentResponseConstructor } from '../../type/web-payment-response-constructor.type';
+import type { Maybe } from '@rnw-community/shared';
+
+
+export const { PaymentResponse }: { PaymentResponse: Maybe<WebPaymentResponseConstructor> | undefined } =
+    typeof window === 'undefined' ? { PaymentResponse: null } : window;