+++ packages/react-native-payments/src/type/web-payment-request-constructor.type.ts
@@ -1,0 +1,1 @@
+export type WebPaymentRequestConstructor = typeof globalThis.PaymentRequest;