+++ packages/shared/src/index.ts
@@ -20,6 +20,7 @@ export { isRecord } from './type-guard/generic/is-record/is-record';
 export { isString } from './type-guard/string/is-string/is-string';
 export { isEmptyString } from './type-guard/string/is-empty-string/is-empty-string';
 export { isNotEmptyString } from './type-guard/string/is-not-empty-string/is-not-empty-string';
+export { isDecimalMonetaryValue } from './type-guard/string/is-decimal-monetary-value/is-decimal-monetary-value';
 
 export { isArray } from './type-guard/array/is-array/is-array';
 export { isEmptyArray } from './type-guard/array/is-empty-array/is-empty-array';