+++ packages/react-native-payments-example/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.11.0](https://github.com/rnw-community/rnw-community/compare/v2.10.0...v2.11.0) (2026-08-02)
+
+**Note:** Version bump only for package @rnw-community/react-native-payments-example
+
+# [2.10.0](https://github.com/rnw-community/rnw-community/compare/v2.9.1...v2.10.0) (2026-08-02)
+
+**Note:** Version bump only for package @rnw-community/react-native-payments-example
+
+## [2.9.1](https://github.com/rnw-community/rnw-community/compare/v2.9.0...v2.9.1) (2026-08-02)
+
+**Note:** Version bump only for package @rnw-community/react-native-payments-example
+
+# [2.9.0](https://github.com/rnw-community/rnw-community/compare/v2.8.1...v2.9.0) (2026-08-02)
+
+**Note:** Version bump only for package @rnw-community/react-native-payments-example
+
+## [2.8.1](https://github.com/rnw-community/rnw-community/compare/v2.8.0...v2.8.1) (2026-08-02)
+
+**Note:** Version bump only for package @rnw-community/react-native-payments-example
+
+# [2.8.0](https://github.com/rnw-community/rnw-community/compare/v2.7.3...v2.8.0) (2026-08-02)
+
+**Note:** Version bump only for package @rnw-community/react-native-payments-example
+
+## [2.7.3](https://github.com/rnw-community/rnw-community/compare/v2.7.2...v2.7.3) (2026-08-02)
+
+**Note:** Version bump only for package @rnw-community/react-native-payments-example
+
+## [2.7.2](https://github.com/rnw-community/rnw-community/compare/v2.7.1...v2.7.2) (2026-08-02)
+
+**Note:** Version bump only for package @rnw-community/react-native-payments-example
+
 ## [2.7.1](https://github.com/rnw-community/rnw-community/compare/v2.7.0...v2.7.1) (2026-08-02)
 
 **Note:** Version bump only for package @rnw-community/react-native-payments-example