+++ packages/rxjs-errors/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.9.0](https://github.com/rnw-community/rnw-community/compare/v2.8.1...v2.9.0) (2026-08-02)
+
+**Note:** Version bump only for package @rnw-community/rxjs-errors
+
 # [2.2.0](https://github.com/rnw-community/rnw-community/compare/v2.1.0...v2.2.0) (2026-05-09)
 
 **Note:** Version bump only for package @rnw-community/rxjs-errors