+++ packages/shared/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.9.0](https://github.com/rnw-community/rnw-community/compare/v2.8.1...v2.9.0) (2026-08-02)
+
+### Features
+
+- **shared:** add isDecimalMonetaryValue guard for decimal currency strings ([713ed54](https://github.com/rnw-community/rnw-community/commit/713ed54e785c2b40ff6a75bf64fc18b5b7e19fe0))
+
 # [2.2.0](https://github.com/rnw-community/rnw-community/compare/v2.1.0...v2.2.0) (2026-05-09)
 
 ### Features